refactor: use snake case for components
