+++ site/src/components/common/footer/styles.tsx
