+++ site/src/components/common/no_script/index.tsx
