+++ site/src/features/geolocation/index.ts
@@ -1,5 +1,5 @@
-export { GeolocationButton } from './components/GeolocationButton'
+export { GeolocationButton } from './components/geolocation_button'
 export { useGeolocation } from './hooks/use_geolocation'
 export { getNearbyStations } from './utils/get_nearby_stations'
 
-export type { GeolocationButtonProps } from './components/GeolocationButton'
+export type { GeolocationButtonProps } from './components/geolocation_button'
