+++ site/src/features/search/index.ts
@@ -1,3 +1,3 @@
-export { SearchBar } from './components/SearchBar'
+export { SearchBar } from './components/search_bar'
 
-export type { SearchBarProps } from './components/SearchBar'
+export type { SearchBarProps } from './components/search_bar'
