+++ site/src/features/toast/index.ts
@@ -1,6 +1,6 @@
 export { useToast } from './stores/toast'
 
-export { Provider as ToastProvider } from './components/Provider'
-export { Toast } from './components/Toast'
+export { Provider as ToastProvider } from './components/provider'
+export { Toast } from './components/toast'
 
-export type { ToastProps } from './components/Toast'
+export type { ToastProps } from './components/toast'
