+++ site/src/pages/_document.tsx
@@ -3,7 +3,7 @@ import { Html, Head, Main, NextScript } from 'next/document'
 import constants from 'src/constants'
 
 import { getCssText } from '@junat/design'
-import { NoScript } from '@components/common/NoScript'
+import { NoScript } from '@components/common/no_script'
 import * as styles from '@junat/design/dist/styles'
 
 import translation from '@utils/translate'
