+++ site/src/pages/train/[date]/[trainNumber].tsx
@@ -9,13 +9,13 @@ import dynamic from 'next/dynamic'
 import { useQuery } from '@tanstack/react-query'
 import { fetchSingleTrain } from '@junat/digitraffic'
 
-import Webmanifest from '@components/common/Webmanifest'
-import Header from '@components/common/Header'
+import Webmanifest from '@components/common/webmanifest'
+import Header from '@components/common/header'
 
 import { useLiveTrainSubscription } from '@hooks/use_live_train_subscription'
 import { useStations } from '@hooks/use_stations'
 
-import Page from '@layouts/Page'
+import Page from '@layouts/page'
 
 import { getLocale } from '@utils/get_locale'
 
@@ -23,7 +23,7 @@ import constants from 'src/constants'
 import translate from '@utils/translate'
 import { Code, getTrainType } from '@utils/get_train_type'
 
-const SingleTimetable = dynamic(() => import('@components/timetables/SingleTimetable'))
+const SingleTimetable = dynamic(() => import('@components/timetables/single_timetable'))
 const DefaultError = dynamic(() => import('next/error'))
 
 interface TrainPageProps {
