+++ site/src/stories/__storybook__/components/copyable.tsx
